+++ source/dred/dred_codegen.h
@@ -3,11 +3,11 @@
 // Converts the given buffer to a C-style static const unsigned char array.
 //
 // Free the returned pointer with free().
-char* dred_codegen_buffer_to_c_array(const unsigned char* buffer, unsigned int size, const char* variableName);
+char* dred_codegen_buffer_to_c_array(const unsigned char* buffer, size_t size, const char* variableName);
 
 // Converts the given buffer to a C-style static const char* string.
 //
 // Free the returned pointer with free().
 //
 // If "variableName" is null, the variable declaration will be excluded.
-char* dred_codegen_buffer_to_c_string(const unsigned char* buffer, unsigned int size, const char* variableName);
+char* dred_codegen_buffer_to_c_string(const unsigned char* buffer, size_t size, const char* variableName);
